Set permissions on ssh uploaded file
When uploading a single file using 'ssh upload', if the user does not set
permissions, then set them from the permissions on the file to upload.

This fixes a bug where Fuego expects scripts that are uploaded to the
board to be executable, if they were executable on the client.
The error was: bash: ./<filename>: Permission denied.

Signed-off-by: Tim Bird <tim.bird@sony.com>