Add acme-capture.sh utility, for capturing power measurement data
acme-capture.sh provides a command line script which uses ssh
to access the sysfs filesystem on an acme board, to read
power data for a probe.  It creates a small script to grab
the data from sysfs, and uses scp to copy the script to the
acme board, and ssh to execute the script to grab the data.

It is intended to be used for a start-capture operation with a
server implementing the REST API.  It runs until terminated.
To terminate cleanly, send the command a SIGTERM.  If only
one instance is running, this can be done with:
pkill -f acme-capture.sh

Signed-off-by: Tim Bird <tim.bird@sony.com>